pyproject.toml
[tool.poetry]
name = "1"
version = "0.1.0"
description = ""
authors = ["UNGBI78 <mybluewings78@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
aerich = "0.8.1"
tomlkit = "0.13.2"
tortoise-orm = {version = "0.23.0", extras = ["asyncpg"]}
python-dotenv = "^1.1.1"


[tool.poetry.group.dev.dependencies]
ruff = "0.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
